Fibonacci4

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,16 +8,16 @@
 
 int main(void)
 {
-	unsigned long fib1 = 1, fib2 = 2, next;
+	unsigned long long fib1 = 1, fib2 = 2, next;
 	int count;
 
-	printf("%llu, %llu", fib1, fib2);
+	printf("%lu, %lu", (unsigned long)fib1, (unsigned long)fib2);
 
 	for (count = 3; count <= 98; count++)
 	{
 	next = fib1 + fib2;
 
-	printf(", %llu", next);
+	printf(", %lu", next);
 
 	fib1 = fib2;
 	fib2 = next;
